Update README.md
Beautiful Soup: This is a popular library for parsing HTML and XML documents. It creates a parse tree for parsing HTML and XML documents, and automatically converts incoming documents to Unicode and outgoing documents to UTF-8. It works well with poorly designed HTML and has a lot of functions.
Requests: This is a simple and powerful HTTP library that can be used to access web pages. It is often used for making API requests and web-based applications. It is the first step in web scraping, as it retrieves the HTML content of a web page.
Csv : The CSV writer library in Python is a powerful tool used for writing data to CSV files. It allows users to convert objects or arrays into a CSV string or write them directly into a file.